Gmail
Extensive testing revealed that without inlining the color on `<a>`
tags that Gmail will use its own color. It’s always best to inline for
this reason, but as a backup I found that using `.classname a`
overrides Gmail’s default link color even if you don’t inline it. (Of
course this still breaks if images are off and then turned on).